:recycle: resolverのファイルを外に出す
